tweaks to ai & training * fixed agent not able to hold * simulate delay on a line clear * look ahead * play multiple games and take the mean * dynamic mutation * more parameters * using f64's to try and reduce rounding errors...

+++ src/game/ai/stats.rs
@@ -1,31 +1,31 @@
-pub fn std_dev(values: &[f32]) -> f32 {
+pub fn std_dev(values: &[f64]) -> f64 {
     if values.is_empty() {
         // HACK
         return 0.0;
     }
 
-    let mean = values.iter().copied().sum::<f32>() / values.len() as f32;
+    let mean = values.iter().copied().sum::<f64>() / values.len() as f64;
     let variance = values.iter()
         .map(|&value| {
             let diff = value - mean;
             diff * diff
         })
-        .sum::<f32>() / values.len() as f32;
+        .sum::<f64>() / values.len() as f64;
     variance.sqrt()
 }
 
 pub trait StdDev {
-    fn std_dev(&self) -> f32;
+    fn std_dev(&self) -> f64;
 }
 
-impl StdDev for &[f32] {
-    fn std_dev(&self) -> f32 {
+impl StdDev for &[f64] {
+    fn std_dev(&self) -> f64 {
         std_dev(self)
     }
 }
 
-impl StdDev for Vec<f32> {
-    fn std_dev(&self) -> f32 {
+impl StdDev for Vec<f64> {
+    fn std_dev(&self) -> f64 {
         std_dev(self)
     }
 }
